backend/functions/requirements.txt
firebase-functions==0.4.2
firebase-admin==6.7.0
google-genai==1.5.0
python-dotenv==1.0.1